requirements.txt
allure-pytest==2.3.2b1
allure-python-commons==2.3.2b1
alphabet-detector==0.0.7
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.7.0
Cerberus==1.3.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
enum34==1.1.6
Flask==1.0.2
gevent==1.3.1
greenlet==0.4.13
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
locustio==0.8.1
lxml==4.3.3
MarkupSafe==1.0
more-itertools==4.1.0
msgpack-python==0.5.6
mysqlclient==1.3.12
parsel==1.5.1
pluggy==0.6.0
py==1.5.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.8.1
pyOpenSSL==19.0.0
pypiwin32==223
pytest==3.5.0
pywin32==224
pyzmq==17.0.0
queuelib==1.5.0
requests==2.18.4
Scrapy==1.6.0
service-identity==18.1.0
six==1.11.0
Twisted==19.2.0
urllib3==1.22
w3lib==1.20.0
Werkzeug==0.14.1
zope.interface==4.6.0
